api_scraper/requirements.txt
bs4==0.0.1
celery==5.2.7
Django==4.1.1
djangorestframework==3.13.1
lxml==4.9.1
psycopg2-binary==2.9.3
redis==4.3.4
requests==2.28.1
pytelegrambotapi==4.7.0
telebot==0.0.4
pip==22.2.2
beautifulsoup4==4.10.0